logger/vxdraw: Create a logger passthrough interface
The logger passthrough is required to cleanly separate loggers from the
main application and libraries. The main application may use specific
loggers to speed up certain operations, while the library periodically
logs and doesn't really care about that, so now we can clone a logger
and give it a generic message.